Allow custom ports #34
Currently the webpack dev server and nodemon are running on a fixed
port.

This commit allows developers to choose their own port.
If no port was specified and the default one is already in use,
it will automatically find the next free port.

For information on how to set the ports, take a look at the readme.

This fixes #34