[feature] Added pre-payment page
Explain the payment (or credit/debit card verification) process
Before logging in the user to the captive portal and send them
to the payment gateway.

Since the user will have limited time to complete the transaction
to avoid abuses (since they basically get free internet for a few
minutes), we have to explain them what to do in case the timeout
is reached.