[fix/change] Fixed hanging logout + force logout after payment #410
After captive portal logout on PfSense, any outgoing HTTP request fails
for a few seconds.

For this reason we should avoid to lazily import Login, which can
get downloaded again and cause the logout process to hang.

Moreover, for the reasons above, during credit/debit card
payment/verification we cannot automatically log out and log in
the user, therefore we shall force a logout and ask the user
to log in after successful registration and payment.

Some informational toast messages with long autoClose have been
added to deal with the cases in which the wait time may last
for many seconds.

Closes #410